Character sprite fix cp 91 (#47)
* Change Stats operator's cast from uint8_t to uint16_t for exp

* Change gif sprite to unique pointer

* Change the way the duck gains levels

+++ src/Stats/Stats.hpp
@@ -11,13 +11,13 @@ struct Stats {
 	Stats operator+(const Stats& right){
 		return { (uint8_t)(min(max(0, happiness + right.happiness), 100)),
 				 (uint8_t)(min(max(0, oilLevel + right.oilLevel), 100)),
-				 (uint8_t)(min(max(0, experience + right.experience), UINT16_MAX)) };
+				 (uint16_t)(min(max(0, experience + right.experience), UINT16_MAX)) };
 	}
 
 	Stats operator-(const Stats& right){
 		return { (uint8_t)(min(max(0, happiness - right.happiness), 100)),
 				 (uint8_t)(min(max(0, oilLevel - right.oilLevel), 100)),
-				 (uint8_t)(min(max(0, experience - right.experience), UINT16_MAX)) };
+				 (uint16_t)(min(max(0, experience - right.experience), UINT16_MAX)) };
 	}
 
 	Stats& operator+=(const Stats& right){
